+++ src/main/java/dev/hoyin1600p/vault_render_optimization/VaultRenderOptimization.java
@@ -1,6 +1,7 @@
 package dev.hoyin1600p.vault_render_optimization;
 
 import dev.hoyin1600p.vault_render_optimization.client.VaultRenderOptimizationCommand;
+import dev.hoyin1600p.vault_render_optimization.client.create.FlywheelBackendManager;
 import dev.hoyin1600p.vault_render_optimization.client.lighting.DynamicLightEngine;
 import dev.hoyin1600p.vault_render_optimization.client.lighting.DynamicLightResourceLoader;
 import dev.hoyin1600p.vault_render_optimization.config.ClientOptimizationConfig;
@@ -9,6 +10,7 @@ import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
+import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
@@ -33,6 +35,9 @@ public final class VaultRenderOptimization {
         modEventBus.addListener(this::onRegisterClientReloadListeners);
         MinecraftForge.EVENT_BUS.addListener(this::onRegisterClientCommands);
         MinecraftForge.EVENT_BUS.addListener(DynamicLightEngine::onClientTick);
+        if (supportsFlywheelBackendManagement()) {
+            MinecraftForge.EVENT_BUS.addListener(FlywheelBackendManager::onClientTick);
+        }
     }
 
     private void onRegisterClientCommands(RegisterClientCommandsEvent event) {
@@ -41,5 +46,14 @@ public final class VaultRenderOptimization {
 
     private void onRegisterClientReloadListeners(RegisterClientReloadListenersEvent event) {
         event.registerReloadListener(DynamicLightResourceLoader.INSTANCE);
+    }
+
+    private static boolean supportsFlywheelBackendManagement() {
+        if (!ModList.get().isLoaded("create") || !ModList.get().isLoaded("flywheel")) {
+            return false;
+        }
+        return ModList.get().getModContainerById("flywheel")
+                .map(container -> container.getModInfo().getVersion().toString().startsWith("0.6.11"))
+                .orElse(false);
     }
 }